strlen to size t

+++ ft_strlen.c
@@ -6,18 +6,18 @@
 /*   By: pgorner <pgorner@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 10:04:09 by pgorner           #+#    #+#             */
-/*   Updated: 2022/11/02 15:04:22 by pgorner          ###   ########.fr       */
+/*   Updated: 2023/01/06 11:48:28 by pgorner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-/* FUNCTION EXPLANATION: 
+/* FUNCTION EXPLANATION:
 ** Function outputs string length
 */
 #include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
